requirements.txt

Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
# mysqlclient==2.1.1

